backend/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.1
